requirements.txt
requests 
beautifulsoup4 
pandas 
scrapy 
selenium 
webdriver-manager
TIME-python
openai
python-dotenv